+++ PainlessMesh/BasicMessage.cs
@@ -12,15 +12,15 @@ namespace PainlessMesh
 {
     public class GeneralSmarthomeMessage
     {
-        [ nj.JsonProperty("id")]
-        public uint NodeId { get; set; }
+        [nj.JsonProperty("id")]
+        public long NodeId { get; set; }
         [nj.JsonProperty("m"), JsonConverter(typeof(StringEnumConverter))]
         public MessageType MessageType { get; set; }
         [nj.JsonProperty("c"), JsonConverter(typeof(StringEnumConverter))]
         public Command Command { get; set; }
         [ nj.JsonProperty("p")/*, nj.JsonConverter(typeof(SingleOrListConverter<JsonElement>))*/]
         public List<JToken> Parameters { get; set; }
-        public GeneralSmarthomeMessage(uint nodeId, MessageType messageType, Command command, params JToken[] parameters)
+        public GeneralSmarthomeMessage(long nodeId, MessageType messageType, Command command, params JToken[] parameters)
         {
             NodeId = nodeId;
             MessageType = messageType;
@@ -32,7 +32,7 @@ namespace PainlessMesh
         }
 
     }
-   
+
 
     public class SubJsonSerializer : JsonConverter
     {
@@ -42,7 +42,7 @@ namespace PainlessMesh
         }
         public override object ReadJson(JsonReader reader, Type objectType, object existingValue, JsonSerializer serializer)
         {
-            var res = new Dictionary<uint, Sub>();
+            var res = new Dictionary<long, Sub>();
 
             res = serializer.Deserialize<Sub[]>(reader).ToDictionary(x => x.NodeId, x => x);
 
@@ -50,7 +50,7 @@ namespace PainlessMesh
         }
         public override void WriteJson(JsonWriter writer, object value, JsonSerializer serializer)
         {
-            var subs = value as Dictionary<uint, Sub>;
+            var subs = value as Dictionary<long, Sub>;
             serializer.Serialize(writer, subs.Values);
         }
     }
@@ -60,11 +60,11 @@ namespace PainlessMesh
     public class Sub : IEquatable<Sub>
     {
         [JsonProperty("nodeId")]
-        public uint NodeId { get; set; }
+        public long NodeId { get; set; }
 
         [JsonConverter(typeof(SubJsonSerializer))]
         [JsonProperty("subs")]
-        public Dictionary<uint, Sub> Subs { get; set; }
+        public Dictionary<long, Sub> Subs { get; set; }
 
         public override bool Equals(object obj) => Equals(obj as Sub);
         public bool Equals(Sub other) => other != null && NodeId == other.NodeId;
